pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regkb"
version = "1.0.0"
description = "Regulatory Knowledge Base - Local document management with natural language search"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "DLSC", email = "info@dlsc.ie"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Indexing",
]
keywords = ["regulatory", "documents", "knowledge-base", "search", "pdf"]

dependencies = [
    "click>=8.1.0",
    "pdfplumber>=0.10.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "alembic>=1.12.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.0",
    "colorlog>=6.7.0",
    "python-dotenv>=1.0.0",
    "msvc-runtime>=14.0.0; sys_platform == 'win32'",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]
web = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "itsdangerous>=2.1.0",
]
bot = [
    "python-telegram-bot[job-queue]>=21.0",
    "apscheduler>=3.10.0,<4.0",
    "feedparser>=6.0.0",
]

[project.scripts]
regkb = "regkb.cli:main"

[project.urls]
Homepage = "https://github.com/dlsc/regulatory-kb"
Documentation = "https://github.com/dlsc/regulatory-kb#readme"
Repository = "https://github.com/dlsc/regulatory-kb"

[tool.setuptools.packages.find]
where = ["scripts"]

[tool.setuptools.package-dir]
"" = "scripts"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"scripts/regkb/cli.py" = ["E402"]
"scripts/regkb/web/main.py" = ["E402"]
"scripts/regkb/web/routes/*.py" = ["B008"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
